python projects /Facial_recognition_system/requirements.txt
opencv-python>=4.8.0
face-recognition>=1.3.0
dlib>=19.24.0
tensorflow>=2.13.0
Pillow>=10.0.0
numpy>=1.24.0
flask>=3.0.0
flask-socketio>=5.3.0
